% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimalTrackR-package.R
\docType{package}
\name{AnimalTrackR-package}
\alias{AnimalTrackR}
\alias{AnimalTrackR-package}
\title{AnimalTrackR: YOLO-based tracking of animals in video for behavioural analysis.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

`AnimalTrackR` provides a simple interface to YOLO11 for use by researchers to track the movement of animals recorded in video footage. It supports the training, development and deployment of detection models for custom datasets. The tracks can be used for high-resolution. long-term behavioural analysis.
}
\author{
\strong{Maintainer}: Mario Lambrette \email{ml673@exeter.ac.uk}

}
\keyword{internal}
