% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackR_env.R
\name{check_gpu}
\alias{check_gpu}
\title{Check if GPU is available}
\usage{
check_gpu(envname = "animaltrackr")
}
\arguments{
\item{envname}{Name of the conda environment in which to check the availability
of a GPU}
}
\value{
Boolean indicating whether GPU is available.
}
\description{
Checks wheter a named conda environment ("animaltrackr" by default) is
correctly configured for GPU acceleration. If you have a CUDA GPU enabled
system you can set the envrionment up to use it properly with
\code{\link[=create_TrackR_env]{create_TrackR_env()}} for a new environment or \code{\link[=configure_cuda]{configure_cuda()}} for
an existing environment.
}
\examples{
\dontrun{
library(AnimalTrackR)
check_gpu()
}

}
\seealso{
\code{\link[=create_TrackR_env]{create_TrackR_env()}} \code{\link[=configure_cuda]{configure_cuda()}}
}
