% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountAnnotations.R
\name{CountImages}
\alias{CountImages}
\title{Count the number of annotated images in a project.}
\usage{
CountImages(project = get_Project())
}
\arguments{
\item{project}{path to a TrackR project directory}
}
\value{
a ggplot object showing the number of images in each set in the specified project
}
\description{
\code{CountImages()} counts the number of annotated images that have been exported
to the YOLO directory of a TrackR project. The default behaviour will count
images in the current active project, found using \code{get-Project()}, though the
path to any TrackR project can be passed to the \code{project} argument instead.

The function returns a bar plot of the number of images in each set  required
for model development (Train, Test and Val).
}
\examples{
\dontrun{
# Count images in the currently active project
CountImages()

# Count images in another project
CountImages(project = "path/to/project/root")
}

}
