% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TrackR_env.R
\name{create_TrackR_env}
\alias{create_TrackR_env}
\title{Create the necessary conda environment for using \code{AnimalTrackR}}
\usage{
create_TrackR_env(envname = "animaltrackr")
}
\arguments{
\item{envname}{Character string. The name of the virtual environment to create}
}
\value{
Invisibly returns TRUE upon successful environment creation and activation.
}
\description{
\code{create_TrackR_env()} creates a python 3.9 conda environment and installs a number
of required packages within it. For the majority of users it only needs to be run once
on the first usage of \code{AnimalTrackR}
}
\details{
The majority of users should run \code{create_TrackR_env()} once when they first use the
\code{AnimalTrackR} package with the default arguments. This will create a conda
environment named "animaltrackr". Each subsequent call to \code{library(AnimalTrackR)}
will automatically connect the user's R session to this conda environment.

The creation of the conda environmen trelies on users having \href{https://docs.anaconda.com/miniconda/}{miniconda}
installed on their machine. If no miniconda installation can be found the
function will prompt users to install miniconda. Internally, accepting this
prompt will call \code{\link[reticulate:install_miniconda]{reticulate::install_miniconda()}} and intall miniconda in the
default locations on the users machine. If users prefer to install their own
version of miniconda they can do so at the link above.

Some users may wish to modify the default behaviour and create an environment
with a different name. The \code{envname} parameter allows this, though users must note
that subsequent calls to \code{library(AnimalTrackR)} will not automatically connect to
the custom environment. After calling \code{library(AnimalTrackR)}, users will need to
use the \code{\link[=set_TrackR_condaenv]{set_TrackR_condaenv()}} function to connect to their custom environment.
}
\examples{
\dontrun{
# Default behaviour (reccomended)
library(AnimalTrackR)
create_TrackR_env() # Not run again on this machine

# Custom environments (Not reccomended)
# On first usage of AnimalTrackR:
library(AnimalTrackR)
create_TrackR_env(envname = 'somename')

# On subsequent usages of AnimalTrackR:
library(AnimalTrackR)
TrackR_condaenv(envname = 'somename')

}

}
\seealso{
\code{\link[reticulate:install_miniconda]{reticulate::install_miniconda()}}, \code{\link[=set_TrackR_condaenv]{set_TrackR_condaenv()}}
}
