% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{demo_run}
\alias{demo_run}
\title{Run an AnimalTrackR detection model in demo mode}
\usage{
demo_run(video, model = NULL, savepath = NULL, gopro = FALSE)
}
\arguments{
\item{video}{Path to video clip to use for detections}

\item{model}{(optional) Name of the TrackR model to use for detections, or a full path to yolov8-11 model weights. NULL, the default, uses the best weights from the first model in the YOLO/models/ directory}

\item{savepath}{Path to a directory into which to save the video detections. NULL, the default, saves the video into the 'Demos' directory of the active TrackR project.}

\item{gopro}{T/F. The default is FALSE but this must be set to TRUE if footage was recorded on a gopro. See details for more information.}
}
\value{
invisibly returns TRUE upon completion
}
\description{
This function allows use to run a detection model on clips of their footage
and visualise the bounding box outputs. This is useful for model validation.
}
\details{
This is a useful function for running detections on example clips to
visualise bounding boxes and visually verify their accuracy. However, it
should not be user's primary method for validating models, for this users
should also look at the
\subsection{Important for GoPro users}{

\strong{\code{gopro} details:} Modern GoPro footage is unfortunately incompatible with
this software (see \href{https://stackoverflow.com/questions/78039408/cv2-ffmpeg-grabframe-packet-read-max-attempts-exceeded-error-after-exactly-rea}{here} for details).
If you try to run a detection model on GoPro footage you may get this error:
\verb{[ WARN:0@2245.518] global cap_ffmpeg_impl.hpp:1595 grabFrame packet read max attempts exceeded}.

If this happens, set the GoPro parameter to TRUE. This reformats the footage
to remove the audio stream. so that it can be processed by the software. The
reformatted file is stored in a temporary location and the original video
file you provide is not modified. This reformatting slows the overall
processing speed. unfortunately this is currently unavoidable, but we are
monitoring updates this software's underlying computer vision framework
(\code{opencv}) for a more comprehensive solution.

Note that this reformatting requires an additional software library, 'FFMPEG'.
This is a free video processing library that can be installed \href{https://ffmpeg.org/download.html}{here}
}
}
\examples{
\dontrun{
  demo_run("path/to/short_clip.mp4")
}

}
