% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HMM.R
\name{.calc_dets_frame_rate}
\alias{.calc_dets_frame_rate}
\title{INTERNAL Calculate median frame rate from frame numbers and timestamps in detection dataframe}
\usage{
.calc_dets_frame_rate(frame, timestamp = NULL, frame_col = NULL)
}
\arguments{
\item{frame}{Either a numeric vector of frame numbers or a dataframe
containing the frame column.}

\item{timestamp}{If \code{frame} is a vector, a numeric vector of timestamps
in milliseconds. If \code{frame} is a dataframe, this argument
should be the name of the timestamp column (string).}

\item{frame_col}{If \code{frame} is a dataframe, the name of the frame column (string).}
}
\value{
Median frame rate (frames per second)
}
\description{
INTERNAL Calculate median frame rate from frame numbers and timestamps in detection dataframe
}
\examples{
\dontrun{
  df <- data.frame(frame = 1:10, timestamp = seq(0, 900, by = 100))
  .calc_dets_frame_rate(df, "frame", "timestamp")
}


}
