% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HMM.R
\name{.check_detections}
\alias{.check_detections}
\title{INTERNAL Check the format of detections dataframe}
\usage{
.check_detections(detections_df)
}
\arguments{
\item{detections_df}{A dataframe of model detections, read from a detections csv file}
}
\value{
T/F depending on whether the detections are formatted correctly
}
\description{
INTERNAL Check the format of detections dataframe
}
