% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HMM.R
\name{.downsample_to_fps}
\alias{.downsample_to_fps}
\title{INTERNAL Downsample detections to a target frame rate}
\usage{
.downsample_to_fps(df, frame_col, dets_fps, state_fps)
}
\arguments{
\item{df}{Dataframe containing detection data}

\item{frame_col}{Name of frame number column (string)}

\item{dets_fps}{Current frame rate of detections}

\item{state_fps}{Desired frame rate}
}
\value{
Downsampled dataframe
}
\description{
INTERNAL Downsample detections to a target frame rate
}
