% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HMM.R
\name{.read_detections}
\alias{.read_detections}
\title{INTERNAL Read a detections csv file in as a dataframe}
\usage{
.read_detections(detections_path)
}
\arguments{
\item{detections_path}{Path to csv of model detections}
}
\value{
A dataframe of model detections
}
\description{
INTERNAL Read a detections csv file in as a dataframe
}
