% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HMM.R
\name{fit_HMM}
\alias{fit_HMM}
\title{Fit a HMM to classify behavioural state of AnimalTrackR detections}
\usage{
fit_HMM(detections, state_fps = 3, overwrite = F)
}
\arguments{
\item{detections}{path to an AnimalTrackR detection file. These are the csv
files stored in the Detections/ directory of an AnimalTrackR project.}

\item{state_fps}{The frame rate at which behavioural classification should
occur. The default is 3. Some applications where behaviour is more finely
resolved than this may benefit from a higher classification frame rate but
in most cases this will add unnecessary noise.}

\item{overwrite}{T/F. If TRUE, existing behavioural state classifications for
the provided detection file will be overwritten.}
}
\value{
a dataframe of detections with a 'state' column showing the
behavioural state. Also saves the updated detections file with a 'State'
column into the original location.
}
\description{
This function takes an AnimalTrackR detection file and fits a Hidden Markov
Model to the bounding coordinates within it to classify behaviour into active
and inactive states. Users supply the path to a detection file and a desired
resolution for behavioural state classification. The updated detection file
with behavioural states supplied in the "State" column is then saved in the
original location.

To update a detection file where behavioural states have already been
calculated, set the \code{overwrite} parameter to TRUE. Otherwise, the function
will not refit states and will throw a warning.
}
\examples{
\dontrun{
 fit_HMM(
   detections = "Detections/Test_detections",
   state_fps = 3
 )
}

}
