% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HMM.R
\name{get_detections_fps}
\alias{get_detections_fps}
\title{Calculate the frame rate of a detections file}
\usage{
get_detections_fps(detections)
}
\arguments{
\item{detections}{Either a file path to an AnimalTrackR detection file or a dataframe of AnimalTrackR detections in the same format as the csv files.}
}
\value{
the FPS of recorded detections in the provided detection file
}
\description{
Calculate the frame rate of a detections file
}
\examples{
\dontrun{
# From file
get_detection_fps("Detections/test_detections.csv")

# From dataframe
detections <- data.table::fread("Detections/test_detections.csv")
get_detections_fps(detections)
}
}
