% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_path.R
\name{get_Project}
\alias{get_Project}
\title{Get the current project root directory}
\usage{
get_Project()
}
\value{
File path to the active project root directory
}
\description{
Allows the user to check the current AnimalTackR project root directory
}
\examples{
\dontrun{
get_Project()
}
}
