% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_path.R
\name{init_Project}
\alias{init_Project}
\title{Initiate a new AnimalTrackR project}
\usage{
init_Project(path = "Project")
}
\arguments{
\item{path}{File path (can be an absolute or relative path) to the root of the
project to initialise. The final part of the file path will be the name of
the project root and therefore the project name so it should be chosen with
some care and be unique among the user's AnimalTrackR projects}
}
\value{
Returns a message in the console confirming successful initialisation
}
\description{
AnimalTrackR relies on a specific project folder structure. Many of its functions
will act on this folder structure implicitly to handle tasks such as dividing
annotated images into training, testing and validation groups. \code{init_Project()}
is a helper function to allow users to initialise a new project and begin
model development.

The function use the path specified as the root to create a full project
folder structure. More information on this structure can be found \href{https://github.com/mariolambrette/AnimalTrackR?tab=readme-ov-file#file-structure}{here}.

A newly initiated project will automatically be set as the current active project.
To change back to a previous project use \code{\link{set_Project}()}
}
\examples{
\dontrun{
# Creates a folder in the current working directory called 'AnimalTrackR-projects'
# then creates the Project folder within that under the 'Project1' root folder

init_project("AnimalTrackR-projects/Project1")
}

}
\seealso{
\code{\link[=set_Project]{set_Project()}}
}
