% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_TrackR.R
\name{install_TrackR}
\alias{install_TrackR}
\title{Create the necessary environment for using \code{AnimalTrackR}}
\usage{
install_TrackR(
  envname = "animaltrackr",
  new_env = identical(envname, "animaltrackr")
)
}
\arguments{
\item{envname}{Character string. The name of the virtual environment to create}

\item{new_env}{Boolean. Whether or not the envname has been changed from the default 'animaltrackr'}
}
\value{
Returns a message in the console confirming successful installation
}
\description{
\code{install_TrackR()} creates a python 3.9 virtual environment and installs a number
of required packages within it. For the majority of users it only needs to be run once
on the first usage of \code{AnimalTrackR}
}
\details{
The majority of users should run \code{install_TrackR()} once when they first use the
\code{AnimalTrackR} package with the default arguments. This will create a virtual
environment named 'animaltrackr'. Each subsequent call to \code{library(AnimalTrackR)}
will automatically connect the user's R session to this virtual environment.

Some users may wish to modify the default behaviour and create an environment
with a different name. The \code{envname} parameter allows this, though users must note
that subsequent calls to \code{library(AnimalTrackR)} will not automatically connect to
the custom environment. Users will need to use the \code{\link[=TrackR_virtualenv]{TrackR_virtualenv()}} function
to connect to their custom environment.
}
\examples{
\dontrun{
# Default behaviour (reccomended)
library(AnimalTrackR)
install_TrackR() # Not run again on this machine

# Custom environments (Not reccomended)
# On first usage of AnimalTrackR:
library(AnimalTrackR)
install_TrackR(envname = 'somename')

# On subsequent usages of AnimalTrackR:
library(AnimalTrackR)
TrackR_virtualenv(envname = 'somename')

}

}
