% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_train_config}
\alias{restore_train_config}
\title{Restore default training hyperparameter configuration file}
\usage{
restore_train_config(project = get_Project())
}
\arguments{
\item{project}{Path to a TrackR project. By default the active project is
retrieved with \code{get_Project()}}
}
\value{
invisibly return TRUE upon completion
}
\description{
This function is used to restore the default training hyperparameter
configuration file after it has been deleted or modified
}
\examples{
\dontrun{
library(AnimaltrackR)
set_Project("path/to/TrackR_project")
restore_train_config()
}
}
