% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzDEV_ModelRun.R
\name{run_Model}
\alias{run_Model}
\title{Run AnimalTrackR detection model}
\usage{
run_Model(
  video,
  detections_path = NULL,
  model = list.files(file.path(get_Project(), "YOLO", "models")[1]),
  save_vid = F,
  save_path = NULL
)
}
\arguments{
\item{video}{Path to the video on which to run inference. This can either be a path to a single video file or a path to a directory containing chapterised video.}

\item{detections_path}{File path at which to save the detections. This is an optional parameter. By default the detections will be saved at "videoPath/videoName.csv".}

\item{model}{The detection model to use for detections. If you have only trained one model in this project you can leave the default behaviour, which uses the first model that appears in the "YOLO/models" directory in the project folder. Otherwise, the name of the desired model folder should be provided. See details for more information.}

\item{save_vid}{Boolean. Defaults to \code{FALSE}, if set to \code{TRUE} a copy of the video is saved with predicted bounding boxes plotted onto it. This is useful for validation purposes.}

\item{save_path}{The path at which to save the video with overlaid detections. Only used if \code{save_vid} is \code{TRUE}.}
}
\value{
Returns the filepath to the detections file
}
\description{
Run AnimalTrackR detection model
}
\details{
\subsection{Add info on specifiying model directory}{
}

\subsection{Add recommendation for saving detections in project directory}{
}
}
\examples{
\dontrun{
 run_model(
   video = "path/to/video.mp4",
 )
}


}
