% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveAnnotations.R
\name{save_annotations}
\alias{save_annotations}
\title{Save annotations into project directory}
\usage{
save_annotations(csv)
}
\arguments{
\item{csv}{path to the csv file exported from \href{https://www.makesense.ai/}{Make Sense}.}
}
\value{
invisibly return TRUE
}
\description{
\code{save_annotations()} takes a csv file of bounding box annotations created with
\href{https://www.makesense.ai/}{Make Sense} and converts them into YOLO format and
saves the relevant annotation files and images into the project directory.
It should be used after completing each annotating session.

As well as saving the completed annotations the function will remove annotated
images from the 'ToAnnotate' folder, ensuring that users do not annotate the
same image twice if they are completing annotations over multiple sessions.
}
\examples{
\dontrun{
# After creating annotations in Make Sense, exporting them to a csv file and
# saving that file to "some/path/annotations.csv":
save_annotations(csv = "some/path/annotation.csv")
}

}
