% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_path.R
\name{set_Project}
\alias{set_Project}
\title{Set TrackR project}
\usage{
set_Project(path)
}
\arguments{
\item{path}{directory path to the desired project root folder}
}
\value{
Invisibly returns TRUE if the project path is set correctly.
}
\description{
AnimalTrackR uses a \href{https://github.com/mariolambrette/AnimalTrackR?tab=readme-ov-file#file-structure}{project folder structure}.
\code{\link{set_Project}()} is a helper function to allow the user to define the root folder
of the project they are working on.
}
\examples{
\dontrun{
set_Project(".../AnimalTrackR-projects/project1")
}
}
