% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackR_env.R
\name{set_TrackR_condaenv}
\alias{set_TrackR_condaenv}
\title{Set the virtual environment for the current R session}
\usage{
set_TrackR_condaenv(envname)
}
\arguments{
\item{envname}{Character string. Name of conda environemnt to connect to.}
}
\value{
Invisibly returns TRUE
}
\description{
For users who have created custom environments or use reticulate in other
contexts frequently this function provides a simple way to set the path to
the correct virtual environment for \code{AnimalTrackR} to run.

Note that the named environment must be a python 3.9 environment with the correct
dependencies installed or the function gives an error.
}
\examples{
\dontrun{
set_TrackR_condaenv('animaltrackr')
}
}
