% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackR_env.R
\name{set_TrackR_env}
\alias{set_TrackR_env}
\title{Set the conda environment for the current session}
\usage{
set_TrackR_env(envname = "animaltrackr")
}
\arguments{
\item{envname}{Character string. Name of conda environemnt to connect to.}
}
\value{
Invisibly returns TRUE
}
\description{
For users who have created custom environments or use reticulate in other
contexts frequently this function provides a simple way to set the path to
the correct conda environment for \code{AnimalTrackR} to run.

Note that the named environment must be a python 3.11 environment with the correct
dependencies installed or the function gives an error. We reccomend you use
\code{\link[=create_TrackR_env]{create_TrackR_env()}} to create the environment to ensure this is the case
}
\examples{
\dontrun{
set_TrackR_env('animaltrackr')
}

}
\seealso{
\code{\link[=create_TrackR_env]{create_TrackR_env()}}
}
