% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_Model}
\alias{train_Model}
\title{Train a YOLOv11s model with data held in a TrackR project directory}
\usage{
train_Model(
  model_name,
  project = get_Project(),
  config = NULL,
  target_only = FALSE
)
}
\arguments{
\item{model_name}{Character string indicating the name of the model to train
(used to distinguish between multiple iterations during development)}

\item{project}{The path to the TrackR project directory. By default this is
supplied by \code{get_Project()}, most users will never need to change this.}

\item{config}{\code{NULL} or the path to a YOLO dataset configuration yaml file.
The default (\code{NULL}) automatically creates a configuration file using the
TrackR project architecture. Most users will never need to change this.}

\item{target_only}{Boolean indicating whether non target annotated objects
should be included in model training. This refers to the 'ZZZ' class which
contains non-target but visually similar objects. There may be cases where
including these in modeltraining actually hampers performance (e.g. where
there are not very many of these objects). This parameters allows those
annotations to be excluded from model training.}
}
\value{
invisibly returns TRUE when training is complete. The model training
results will be available to view at 'YOLO/model/\{model_name\}'
}
\description{
This function takes the annotated data in a TrackR project directory and uses
it to train a YOLO object detection model. This model can then be evaluated
and used to track animals in a broader dataset.
}
\details{
The base weights used for training are the default \href{https://docs.ultralytics.com/models/yolo11/#supported-tasks-and-modes}{YOLO11s detection weights}.
Training is configured to run for 200 epochs, with an image size of 640 and a
batch size of 16. In almost all cases users will not need to worry about
these parameters, with sufficient training data they should enable up to 100\%
accurate detection models to be trained.

For advanced users, or if you have a specific requirement for your model,
these training hyperparamters can be modified. The training hyperparamters
are specified in the 'train_config.yaml' file stored in the 'YOLO/configs/'
directory of a TrackR project. users can navigate to this file and modify the
already specified hyperparameters or add any other hyperparameters they wish.
This should be done with caution as it could significantly harm training
results but may be essential in some cases. For a full description of
possible hyperparameters see the \href{https://docs.ultralytics.com/modes/train/#train-settings}{ultralytics documentation}
}
\examples{
\dontrun{

# After images have been annotated, set the active project
library(AnimaltrackR)
set_Project("path/to/TackR_project")

# Train model on annotations
train_Model("model1")

# Training results will be available in the 'YOLO/model/model1' directory
}

}
