% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimalTrackR-package.R
\docType{package}
\name{AnimalTrackR-package}
\alias{AnimalTrackR}
\alias{AnimalTrackR-package}
\title{AnimalTrackR: YOLO-based tracking of animals in captivity for behavioural analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mario Lambrette \email{ml673@exeter.ac.uk}

}
\keyword{internal}
