% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackR_env.R
\name{check_TrackR_env}
\alias{check_TrackR_env}
\title{Check the current TrackR environment}
\usage{
check_TrackR_env()
}
\value{
TRUE if the environment meets the requirements, or FALSE if not.
}
\description{
Checks the active conda environment meets the requirements to run
AnimalTrackR.
}
\examples{
\dontrun{
check_TrackR_env()
}


}
